% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DefPos}
\alias{DefPos}
\title{Function that deals with negative eigen values in a matrix not positive definite}
\usage{
DefPos(X)
}
\arguments{
\item{M}{the matrix or the vector}
}
\value{
the new matrix M
}
\description{
Function that deals with negative eigen values in a matrix not positive definite
}
