% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\docType{data}
\name{estim.class}
\alias{estim.class}
\title{A Reference Class to generate different calibration methods after generating a model from
model.class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
estim.class
}
\description{
A Reference Class to generate different calibration methods after generating a model from
model.class
}
\examples{

}
\keyword{datasets}
