% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\docType{data}
\name{forecast.class}
\alias{forecast.class}
\title{A Reference Class to generates differents \code{forecast.class} objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
forecast.class
}
\description{
See the function \code{\link{forecast}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link{forecast}}... Other methods
should not be called as they are designed to be used during the calibration process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{modelfit}}{a \code{\link{calibrate.class}} object generated by \code{\link{calibrate}}}

\item{\code{md.new}}{the new model created with the MAP estimator}

\item{\code{x.new}}{a new data set (according to the data set used for calibration)}
}}

\keyword{datasets}
