% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%<\%}
\alias{\%<\%}
\title{Operator to define active bindings variables}
\usage{
md \%<\% param
}
\arguments{
\item{md}{a statistical model defined by the function \code{\link{model}}}

\item{param}{a \code{list} of parameter values}
}
\value{
a \code{\link{model.class}} parametrized.
}
\description{
Operator to define active bindings variables
}
