% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\docType{data}
\name{Kernel.class}
\alias{Kernel.class}
\title{A Reference Class to generates differents model objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Kernel.class
}
\description{
See the function \code{\link{model}} which produces an instance of this class
}
\keyword{datasets}
