% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kernelFun}
\alias{kernelFun}
\title{Generates \code{\link{Kernel.class}} covariances matrices}
\usage{
kernelFun(X, var, theta, kernel.type)
}
\arguments{
\item{X}{data}

\item{newdata}{newdata for the prediction}
}
\value{
return a \code{\link{predict.class}} object with two main methods
}
\description{
\code{Kernel} is a function that allows us to generate covariances matrices from data
}
\details{
The realized estimation is realized similarly as it is defined in [1]
}
\examples{
X <- cbind(seq(0,10,length.out=10),seq(8,20,length.out=10))
var <- 2
theta <- 0.1
test <- kernelFun(X,var,theta,kernel.type="gauss")

}
\seealso{
\code{\link{model.class}}, \code{\link{prior.class}}, \code{\link{esim.class}}
}
\author{
M. Carmassi
}
