% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\docType{data}
\name{predict.class}
\alias{predict.class}
\title{Function that realize the prediction after calibration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
predict.class
}
\arguments{
\item{code}{the computational code (function of X and theta)}

\item{X}{the matrix of forced variables}

\item{Yexp}{the vector of the experiments}

\item{model}{string of the model chosen ("model1","model2")}
}
\value{
A function f(theta) is return
}
\description{
Function that realize the prediction after calibration
}
\examples{
X <- cbind(runif(3),runif(3))
code <- function(X,theta)
{
 return(X[,1]+theta*X[,2])
}
Yexp <- runif(3)
test <- model(code,X,Yexp,"model1")
test(2,3)
}
\keyword{datasets}
