% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\docType{data}
\name{prediction.class}
\alias{prediction.class}
\title{Function that realize the prediction after calibration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
prediction.class
}
\description{
Function that realize the prediction after calibration
}
\keyword{datasets}
