% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prediction}
\alias{prediction}
\title{Generates \code{\link{predict.class}} objects}
\usage{
prediction(est, x.new)
}
\arguments{
\item{est}{a \code{\link{estim.class}} object}

\item{x.new}{newdata for the prediction}
}
\value{
lala for the moment
}
\description{
\code{predict} is a function that allows us to generate a class in which the estimation is
done
}
\details{
The realized estimation is realized similarly as it is defined in [1]
}
\examples{
####### For the first model
### The data set
X <- seq(0,1,length.out=100)
### The code to calibrate
code <- function(X,theta)
{
  return((6*X-2)^2*sin(theta*X-4))
}
### Simulated data
Yr   <- code(X,11)
Yexp <- Yr+rnorm(100,0,0.1)
### Definition of the nature of the priors
type.prior=c("gaussian","gamma")
### Definition of the prior hyperparameters
opt.prior=list(c(11,3),c(2,0.1))
### Definition of the estimation option
opt.estim=list(Ngibbs=400,Nmh=1000,thetaInit=c(11,0.1),k=c(5e-4,5e-4),sig=diag(2))

test <- estim(code,X,Yr,Yexp,model="model1",type.prior,log=TRUE,opt.prior,opt.estim)
test$plot()

}
\seealso{
\code{\link{model.class}}, \code{\link{prior.class}}
}
\author{
M. Carmassi
}
