% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{data}
\name{prior.class}
\alias{prior.class}
\title{A Reference Class to generates differents \code{\link{prior.class}} objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
prior.class
}
\description{
See the function \code{\link{prior}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation \code{\link{prior}}. Other methods
 should not be called as they are designed to be used during the optimization process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{type.prior}}{of the selected prior}

\item{\code{opt.prior}}{the charasteristics of the selected prior}

\item{\code{log}}{if we want the log result or not}
}}

\keyword{datasets}
