% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unscale.matrix.diag}
\alias{unscale.matrix.diag}
\title{Funcion which unscale only the diagonal component of a matrix}
\usage{
unscale.matrix.diag(M, binf, bsup)
}
\arguments{
\item{M}{the matrix}

\item{binf}{the lower bound}

\item{bsup}{the upper bound}
}
\value{
the normalized diagonal
}
\description{
Funcion which unscale only the diagonal component of a matrix
}
\examples{
X <- ones(3)*runif(3)
Y <-unscale.matrix.diag(X,c(10,10,1O),c(15,15,15))
print(Y)
}
