% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unscale}
\alias{unscale}
\title{Function which unscale un matrix or a vector}
\usage{
unscale(M, binf, bsup, diag = FALSE, sym = FALSE)
}
\arguments{
\item{M}{the matrix or the vector}

\item{binf}{the lower bound}

\item{bsup}{the upper bound}

\item{diag}{default value False if we want to unscale the whole matrix}

\item{sym}{default value False if we do not have a symmetric matrix}
}
\value{
the unscaled vector or matrix
}
\description{
Function which unscale un matrix or a vector
}
\examples{
\dontrun{
X <- diag(3)*runif(3)
Y <- unscale(X,rep(10,3),rep(15,3))
print(Y)
}
}
