% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate.class}
\alias{calibrate.class}
\title{A Reference Class to generates different \code{calibrate.class} objects}
\description{
See the function \code{\link{calibrate}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link{calibrate}}... Other methods
should not be called as they are designed to be used during the calibration process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{md}}{a \code{\link{model.class}} object generated by \code{\link{model}}}

\item{\code{pr}}{a \code{\link{prior.class}} object generated by \code{\link{prior}}}

\item{\code{opt.estim}}{list of estimation options (see \code{\link{calibrate}} for details)}

\item{\code{opt.valid}}{list of cross validation options (see \code{\link{calibrate}} for details)}

\item{\code{logPost}}{the log posterior}

\item{\code{mcmc}}{a \code{coda} variable of the generated chains}

\item{\code{output}}{list of several chains and acceptation ratios}

\item{\code{onlyCV}}{activates only the cross validation}

\item{\code{errorCV}}{output of the CV errors}

\item{\code{n.cores}}{number cores available}

\item{\code{binf}}{the lower bound of the parameters for the DOE}

\item{\code{bsup}}{the upper bound of the parameters for the DOE}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calibrate.class-new}{\code{calibrate.class$new()}}
\item \href{#method-calibrate.class-calibration}{\code{calibrate.class$calibration()}}
\item \href{#method-calibrate.class-CV}{\code{calibrate.class$CV()}}
\item \href{#method-calibrate.class-calibrationCV}{\code{calibrate.class$calibrationCV()}}
\item \href{#method-calibrate.class-plot}{\code{calibrate.class$plot()}}
\item \href{#method-calibrate.class-acf}{\code{calibrate.class$acf()}}
\item \href{#method-calibrate.class-corrPlot}{\code{calibrate.class$corrPlot()}}
\item \href{#method-calibrate.class-mcmcChains}{\code{calibrate.class$mcmcChains()}}
\item \href{#method-calibrate.class-outputPlot}{\code{calibrate.class$outputPlot()}}
\item \href{#method-calibrate.class-print}{\code{calibrate.class$print()}}
\item \href{#method-calibrate.class-clone}{\code{calibrate.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-new"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$new(md = NA, pr = NA, opt.estim = NA, opt.valid = NULL, onlyCV)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-calibration"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-calibration}{}}}
\subsection{Method \code{calibration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$calibration(i = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-CV"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-CV}{}}}
\subsection{Method \code{CV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$CV(i = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-calibrationCV"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-calibrationCV}{}}}
\subsection{Method \code{calibrationCV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$calibrationCV(mdTemp, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-plot"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$plot(x, graph = "all", label = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-acf"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-acf}{}}}
\subsection{Method \code{acf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$acf(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-corrPlot"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-corrPlot}{}}}
\subsection{Method \code{corrPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$corrPlot(i, j)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-mcmcChains"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-mcmcChains}{}}}
\subsection{Method \code{mcmcChains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$mcmcChains(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-outputPlot"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-outputPlot}{}}}
\subsection{Method \code{outputPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$outputPlot(select.X = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-print"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibrate.class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-calibrate.class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrate.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
