% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estim}
\alias{estim}
\title{Generates \code{\link{estim.class}} objects}
\usage{
estim(code, X, Yexp, model = "model1", type.prior, log = FALSE,
  opt.emul = list(p = 1, n.emul = 100, PCA = TRUE, binf = 0, bsup = 1),
  opt.prior)
}
\arguments{
\item{md}{a \code{\link{model.class}} object}

\item{pr}{a \code{\link{prior.class}} object}

\item{x}{data for calibration}
}
\value{
opt list of options for the inference
}
\description{
\code{estim} is a function that allows us to generate a class in which the estimation is
done
}
\details{
The realized estimation is realized similarly as it is defined in [1]
}
\examples{


}
\seealso{
\code{\link{model.class}}, \code{\link{prior.class}}
}
\author{
M. Carmassi
}
