% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estim}
\alias{estim}
\title{Generates \code{\link{estim.class}} objects}
\usage{
estim(code, X, Yexp, model = "model1", type.prior, log = TRUE,
  opt.emul = list(p = 1, n.emul = 100, PCA = TRUE, binf = 0, bsup = 1),
  opt.prior, opt.estim)
}
\arguments{
\item{md}{a \code{\link{model.class}} object}

\item{pr}{a \code{\link{prior.class}} object}

\item{x}{data for calibration}
}
\value{
opt list of options for the inference
}
\description{
\code{estim} is a function that allows us to generate a class in which the estimation is
done
}
\details{
The realized estimation is realized similarly as it is defined in [1]
}
\examples{
### For the first model
X <- cbind(runif(3),runif(3))
code <- function(X,theta)
{
  return(X[,1]+as.vector(theta)*X[,2])
}
Yexp <- runif(3)
type.prior=c("gaussian","gamma")
opt.emul=list(p=1,n.emul=100,PCA=TRUE,binf=0,bsup=1)
opt.prior=list(c(3,1),c(0.2,0.3))
opt.estim=list(Ngibbs=3000,Nmh=10000,thetaInit=c(3,1),k=c(0.1,0.1),sig=diag(2))
test <- estim(code,X,Yexp,model="model1",type.prior,log=TRUE,opt.emul,opt.prior,opt.estim)
plot(density(test$THETA[,1]),type='l',xlim=c(0,4))
plot(density(test$THETA[,2]),type='l')
}
\seealso{
\code{\link{model.class}}, \code{\link{prior.class}}
}
\author{
M. Carmassi
}
