% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.class}
\alias{forecast.class}
\title{A Reference Class to generates differents \code{forecast.class} objects}
\description{
See the function \code{\link{forecast}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link{forecast}}... Other methods
should not be called as they are designed to be used during the calibration process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{modelfit}}{a \code{\link{calibrate.class}} object generated by \code{\link{calibrate}}}

\item{\code{md.new}}{the new model created with the MAP estimator}

\item{\code{x.new}}{a new data set (according to the data set used for calibration)}

\item{\code{y.new.pred}}{a data set containing the forecasted y for the MAP of theta}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-forecast.class-new}{\code{forecast.class$new()}}
\item \href{#method-forecast.class-print}{\code{forecast.class$print()}}
\item \href{#method-forecast.class-plot}{\code{forecast.class$plot()}}
\item \href{#method-forecast.class-clone}{\code{forecast.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forecast.class-new"></a>}}
\if{latex}{\out{\hypertarget{method-forecast.class-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{forecast.class$new(modelfit, md.new, x.new)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forecast.class-print"></a>}}
\if{latex}{\out{\hypertarget{method-forecast.class-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{forecast.class$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forecast.class-plot"></a>}}
\if{latex}{\out{\hypertarget{method-forecast.class-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{forecast.class$plot(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forecast.class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-forecast.class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{forecast.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
