% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel.R
\docType{data}
\name{Kernel.class}
\alias{Kernel.class}
\title{A Reference Class to generates differents model objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Kernel.class
}
\description{
See the function \code{\link{model}} which produces an instance of this class
}
\section{Fields}{

\describe{
\item{\code{code}}{a function which takes in entry X and theta}

\item{\code{X}}{the matrix of the forced variables}

\item{\code{var}}{the variance for the covariance function}

\item{\code{psi}}{the scale vector of correlation length}

\item{\code{n}}{number of experiments}

\item{\code{Kernel.type}}{the chosen form of covariance}

\item{\code{Cov}}{the covariance}
}}

\keyword{datasets}
