% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model.class}
\alias{model.class}
\title{A Reference Class to generates differents model objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model.class
}
\description{
See the function blabla which produces an instance of this class
}
\section{Fields}{

\describe{
\item{\code{code}}{a function which takes in entry X and theta}

\item{\code{X}}{the matrix of the forced variables}

\item{\code{theta}}{the vector of parameters to estimate}

\item{\code{Yexp}}{the experimental output}

\item{\code{n}}{the number of experiments}

\item{\code{model}}{the model choice see documentation}
}}

\seealso{
The function ....
}
\keyword{datasets}
