% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model.class}
\alias{model.class}
\title{A Reference Class to generates differents model objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
model.class
}
\description{
See the function \code{\link{model}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link{model}}... Other methods
should not be called as they are designed to be used during the calibration process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{code}}{a function which takes in entry X and theta}

\item{\code{X}}{the matrix of the forced variables}

\item{\code{Yexp}}{the experimental output}

\item{\code{n}}{the number of experiments}

\item{\code{opt.emul}}{a list of parameter for the surrogate \itemize{
\item{\strong{p}}{ the number of parameter in the model (defaul value 1)}
\item{\strong{n.emul}}{ the number of points for contituing the Design Of Experiments (DOE) (default value 100)}
\item{\strong{type}}{ type of the chosen kernel (value by default "matern5_2") from \code{\link{km}} function}
\item{\strong{binf}{ the lower bound of the parameter vector (default value 0)}}
\item{\strong{bsup}{ the upper bound of the parameter vector (default value 1)}}
\item{\strong{DOE}{ design of experiments for the surrogate (default value NULL)}}
}}

\item{\code{model}}{the model choice (see \code{\link{model}} for more specification).}
}}

\keyword{datasets}
