% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\docType{data}
\name{prediction.class}
\alias{prediction.class}
\title{A Reference Class to generates differents \code{prediction.class} objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
prediction.class
}
\description{
See the function \code{\link{prediction}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link{prediction}}... Other methods
should not be called as they are designed to be used during the calibration process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{modelfit}}{a \code{\link{calibrate.class}} object generated by \code{\link{calibrate}}}

\item{\code{x.new}}{a new data set (according to the data set used for calibration)}

\item{\code{lowerPred}}{the lower bound of the 90\% credibility interval}

\item{\code{upperPred}}{the upper bound of the 90\% credibility interval}

\item{\code{meanPred}}{the predicted value}
}}

\keyword{datasets}
