% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior.class}
\alias{prior.class}
\title{A Reference Class to generate differents \code{\link{prior.class}} objects}
\description{
See the function \code{\link{prior}} which produces an instance of this class
This class comes with a set of methods, some of them being useful for the user:
See the documentation \code{\link{prior}}. Other methods
 should not be called as they are designed to be used during the optimization process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type.prior}}{of the selected prior}

\item{\code{opt.prior}}{the characteristics of the selected prior}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-prior.class-new}{\code{prior.class$new()}}
\item \href{#method-prior.class-centerText}{\code{prior.class$centerText()}}
\item \href{#method-prior.class-clone}{\code{prior.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior.class-new"></a>}}
\if{latex}{\out{\hypertarget{method-prior.class-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prior.class$new(type.prior = NA, opt.prior = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior.class-centerText"></a>}}
\if{latex}{\out{\hypertarget{method-prior.class-centerText}{}}}
\subsection{Method \code{centerText()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prior.class$centerText(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior.class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-prior.class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prior.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
