% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prior}
\alias{prior}
\title{Generates \code{\link{prior.class}} objects}
\usage{
prior(type.prior, opt.prior, log = FALSE)
}
\arguments{
\item{type.prior}{the vector of the prior types selected}

\item{opt.prior}{list of the hyperparameters relatives to the prior selected}
}
\value{
\code{prior} returns a \code{prior.class} object
}
\description{
\code{prior} is a function that allows us to generate one or a list of classes
 in which the charasteristics of each prior are defined
}
\details{
The realized estimation is realized similarly as it is defined in [1]
}
\examples{
### Only one prior is wanted
###### For a Gaussian Prior
foo <- prior(type.prior="gaussian",opt.prior=list(c(0.5,0.001)))
foo$plot()

###### For a Uniform Prior
foo <- prior(type.prior="unif",opt.prior=list(c(0,1)))
foo$plot()

###### For a Gamma Prior
foo <- prior(type.prior="gamma",opt.prior=list(c(5,1)))
foo$plot()

##### For an inverse-Gamma Prior
foo <- prior(type.prior="invGamma",opt.prior=list(c(0.2,0.3)))
foo$plot()

### For several priors
foo <- prior(type.prior=c("gaussian","gamma"),opt.prior=list(c(0.5,0.001),c(5,1)))
grid.arrange(foo$Prior1$plot(),foo$Prior2$plot(),nrow=2)

}
\seealso{
\code{\link{model.class}}, \code{\link{estim.class}}
}
\author{
M. Carmassi
}
