% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialDesign.R
\name{seqDesign.class}
\alias{seqDesign.class}
\title{A Reference Class to generate a better Model2 or Model4 \code{\link{model.class}} objects}
\description{
This class generates a new \code{\link{model.class}} for Model4 and Model2. Based on the previous
estimation of the Gaussian process in the function \code{\link{model}}, the design of experiments previously used
is improved according to [Damblin et al. 2018]. The aim is to reduce the
error produced by the initial estimation of the Gaussian process by fortifying the initial DOE. The method consists
in proposing new points based on the expectancy improvement criterion.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\references{
DAMBLIN, Guillaume, BARBILLON, Pierre, KELLER, Merlin, et al. Adaptive numerical designs for the
 calibration of computer codes. SIAM/ASA Journal on Uncertainty Quantification, 2018, vol. 6, no 1, p. 151-179.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{doe.init}}{the initial DOE used to fit the first Gaussian process}

\item{\code{GP.init}}{the initial Gaussian process generated in \code{\link{model}} function}

\item{\code{GP.new}}{the new Gaussian process fortified with the new design points}

\item{\code{p}}{the number of parameter}

\item{\code{md}}{the initial model}

\item{\code{md.new}}{the new model}

\item{\code{mdfit}}{the initial calibrated model}

\item{\code{mdfit.new}}{the new calibrated model}

\item{\code{X}}{the data set}

\item{\code{m}}{minimum of the sum of squares used in the algorithm}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seqDesign.class-new}{\code{seqDesign.class$new()}}
\item \href{#method-seqDesign.class-Crit}{\code{seqDesign.class$Crit()}}
\item \href{#method-seqDesign.class-fun.new}{\code{seqDesign.class$fun.new()}}
\item \href{#method-seqDesign.class-SS}{\code{seqDesign.class$SS()}}
\item \href{#method-seqDesign.class-EI}{\code{seqDesign.class$EI()}}
\item \href{#method-seqDesign.class-plot}{\code{seqDesign.class$plot()}}
\item \href{#method-seqDesign.class-clone}{\code{seqDesign.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-new"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$new(md = NA, pr = NA, opt.estim = NA, k = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-Crit"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-Crit}{}}}
\subsection{Method \code{Crit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$Crit(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-fun.new"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-fun.new}{}}}
\subsection{Method \code{fun.new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$fun.new(X, theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-SS"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-SS}{}}}
\subsection{Method \code{SS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$SS(theta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-EI"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-EI}{}}}
\subsection{Method \code{EI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$EI(theta, k = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-plot"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$plot(x, graph = "all", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqDesign.class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seqDesign.class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqDesign.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
