% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unscale.matrix.diag}
\alias{unscale.matrix.diag}
\title{Function which unscale only the diagonal component of a matrix}
\usage{
unscale.matrix.diag(M, binf, bsup)
}
\arguments{
\item{M}{the matrix}

\item{binf}{the lower bound}

\item{bsup}{the upper bound}
}
\value{
the normalized diagonal
}
\description{
Function which unscale only the diagonal component of a matrix
}
\examples{
\dontrun{
X <- diag(3)*runif(3)
Y <-unscale.matrix.diag(X,rep(10,3),rep(15,3))
print(Y)
}
}
