% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unscale.vector}
\alias{unscale.vector}
\title{Function which unscale a vector between two bounds}
\usage{
unscale.vector(x, binf, bsup)
}
\arguments{
\item{x}{the vector to unscale}

\item{binf}{the lower bound}

\item{bsup}{the upper bound}
}
\value{
y the vector unscaled
}
\description{
Function which unscale a vector between two bounds
}
\examples{
\dontrun{
X <- runif(3)
Y <-unscale.vector(X,rep(10,3),rep(15,3))
print(Y)
}
}
